core: Separate boot_final from boot_init_primary_late
The SP-s are initialized at the end of the primary_late step of the
boot sequence. This part needs to be separated from the primary_late
step, so pointer authentication can be enabled right before the SP
initialization. This change does not affect the boot sequence,
contains only code structural change.

Signed-off-by: Gabor Toth <gabor.toth2@arm.com>